+++ src/__tests__/client/textSynchronization.test.ts
@@ -1,4 +1,4 @@
-import { Neovim } from '../../neovim'
+import { Neovim } from '@chemzqm/neovim'
 import fs from 'fs'
 import os from 'os'
 import path from 'path'
@@ -164,8 +164,10 @@ describe('TextDocumentSynchronization', () => {
       await nvim.call('setline', [1, 'foo'])
       await doc.synchronize()
       await client.forceDocumentSync()
+      await helper.wait(50)
       await nvim.call('setline', [1, 'bar'])
       await doc.synchronize()
+      await client.forceDocumentSync()
       await helper.waitValue(() => {
         return called
       }, true)