+++ src/__tests__/handler/workspace.test.ts
@@ -1,4 +1,4 @@
-import { Neovim } from '../../neovim'
+import { Neovim } from '@chemzqm/neovim'
 import fs from 'fs'
 import os from 'os'
 import path from 'path'
@@ -184,7 +184,7 @@ describe('Workspace handler', () => {
       await handler.openLocalConfig()
       await nvim.command(`e ${path.join(os.tmpdir(), 't.vim')}`)
       await nvim.command('setf vim')
-      await handler.openLocalConfig()
+      await commands.executeCommand('workspace.openLocalConfig')
     })
 
     it('should open local config', async () => {