+++ src/__tests__/snippets/snippet.test.ts
@@ -1,10 +1,10 @@
-import { Neovim } from '../../neovim'
+import { Neovim } from '@chemzqm/neovim'
 import path from 'path'
 import { CancellationTokenSource } from 'vscode-languageserver-protocol'
 import { Position, Range, TextEdit } from 'vscode-languageserver-types'
 import { URI } from 'vscode-uri'
 import { LinesTextDocument } from '../../model/textdocument'
-import { addPythonTryCatch, convertRegex, executePythonCode, getVariablesCode, UltiSnippetContext } from '../../snippets/eval'
+import { addPythonTryCatch, convertRegex, evalCode, executePythonCode, getVariablesCode, UltiSnippetContext } from '../../snippets/eval'
 import { Placeholder, TextmateSnippet, Variable } from '../../snippets/parser'
 import { checkContentBefore, CocSnippet, comparePlaceholder, getContentBefore, getEndPosition, getParts, normalizeSnippetString, reduceTextEdit, shouldFormat } from '../../snippets/snippet'
 import { padZero, parseComments, parseCommentstring, SnippetVariableResolver } from '../../snippets/variableResolve'
@@ -295,8 +295,14 @@ describe('CocSnippet', () => {
     })
 
     it('should compares placeholders', () => {
-      expect(comparePlaceholder({ primary: false, index: 1 }, { primary: false, index: 0 })).toBe(-1)
-      expect(comparePlaceholder({ primary: true, index: 1 }, { primary: false, index: 1 })).toBe(-1)
+      let arr = [
+        { primary: false, index: 1, nestCount: 2 },
+        { primary: true, index: 2, nestCount: 1 },
+        { primary: false, index: 3, nestCount: 0 },
+      ]
+      arr.sort(comparePlaceholder)
+      let indexes = arr.map(p => p.index)
+      expect(indexes).toEqual([3, 2, 1])
     })
   })
 
@@ -493,12 +499,21 @@ describe('CocSnippet', () => {
     it('should normalizeSnippetString', () => {
       expect(normalizeSnippetString('a\n\n\tb', '  ', {
         insertSpaces: true,
+        trimTrailingWhitespace: true,
         tabSize: 2
       })).toBe('a\n\n    b')
       expect(normalizeSnippetString('a\n\n  b', '\t', {
         insertSpaces: false,
+        trimTrailingWhitespace: true,
         tabSize: 2
       })).toBe('a\n\n\t\tb')
+      let res = normalizeSnippetString('a\n\n\tb', '\t', {
+        insertSpaces: false,
+        trimTrailingWhitespace: false,
+        noExpand: true,
+        tabSize: 2
+      })
+      expect(res).toBe('a\n\t\n\t\tb')
     })
 
     it('should throw for invalid regex', async () => {
@@ -543,6 +558,17 @@ describe('CocSnippet', () => {
       let msg = await nvim.getVar('errmsg')
       expect(msg).toBeDefined()
       expect(msg).toMatch('INVALID_CODE')
+    })
+
+    it('should eval code', async () => {
+      let val = process.env.SHELL
+      process.env.SHELL = '/bin/sh'
+      let res = await evalCode(nvim, 'shell', 'echo foo', '')
+      expect(res).toBe('foo')
+      process.env.SHELL = undefined
+      res = await evalCode(nvim, 'shell', 'echo foo', '')
+      expect(res).toBe('foo')
+      process.env.SHELL = val
     })
 
     it('should parse comments', async () => {