+++ src/completion/source.ts
@@ -1,5 +1,5 @@
 'use strict'
-import { Neovim } from '../neovim'
+import { Neovim } from '@chemzqm/neovim'
 import events from '../events'
 import { defaultValue, disposeAll, getConditionValue, waitImmediate } from '../util'
 import { isFalsyOrEmpty, toArray } from '../util/array'
@@ -97,9 +97,8 @@ export default class Source implements ISource<ExtendedCompleteItem> {
     return this.getConfig('firstMatch', true)
   }
 
-  // exists opitonnal function names for remote source
-  public get optionalFns(): string[] {
-    return this.defaults['optionalFns'] || []
+  public get remoteFns(): string[] {
+    return toArray(this.defaults.remoteFns)
   }
 
   public get shortcut(): string {