+++ typings/index.d.ts
@@ -6242,6 +6242,18 @@ declare module 'coc.nvim' {
      * When getting a text editor's options, this property will always be a boolean (resolved).
      */
     insertSpaces: boolean
+    /**
+     * Trim trailing whitespace on a line.
+     */
+    trimTrailingWhitespace?: boolean
+    /**
+     * Insert a newline character at the end of the file if one does not exist.
+     */
+    insertFinalNewline?: boolean
+    /**
+     * Trim all newlines after the final newline at the end of the file.
+     */
+    trimFinalNewlines?: boolean
   }
 
   /**
@@ -6758,6 +6770,11 @@ declare module 'coc.nvim' {
      * Reload current buffer by `:edit` command.
      */
     export function executeCommand(command: 'workbench.action.reloadWindow'): Promise<void>
+
+    /**
+     * Open user's coc-settings.json configuration file.
+     */
+    export function executeCommand(command: 'workbench.action.openSettingsJson'): Promise<void>
 
     /**
      * Insert snippet at range of current buffer.
@@ -8702,6 +8719,7 @@ declare module 'coc.nvim' {
 
   export namespace workspace {
     export const nvim: Neovim
+    export const isTrusted = true
     /**
      * Current buffer number, could be wrong since vim could not send autocmd as expected.
      *
@@ -8837,6 +8855,14 @@ declare module 'coc.nvim' {
      * @return A path relative to the root or the input.
      */
     export function asRelativePath(pathOrUri: string | Uri, includeWorkspaceFolder?: boolean): string
+
+    /**
+     * Returns converted unix path when the vim is built with win32unix enabled. Original fullpath is returned when the
+     * convert is not necessary.  Only needed when the fullpath is passed vim directly.
+     *
+     * @param fullpath The filepath to fix, only windows absolute filepath is fixed.
+     */
+    export function fixWin32unixFilepath(fullpath: string): string
 
     /**
      * Opens a document. Will return early if this document is already open. Otherwise
@@ -10718,11 +10744,26 @@ declare module 'coc.nvim' {
   }
 
   export interface UltiSnippetOption {
+    /**
+     * Regex text for regex snippet.
+     */
     regex?: string
+    /**
+     * Context code to execute.
+     */
     context?: string
-    noPython?: boolean
-    range?: Range
-    line?: string
+    /**
+     * Do not expand tabs.
+     */
+    noExpand?: boolean
+    /**
+     * Trim all whitespaces from right side of snippet lines.
+     */
+    trimTrailingWhitespace?: boolean
+    /**
+     * Remove whitespace immediately before the cursor at the end of a line before jumping to the next tabstop
+     */
+    removeWhiteSpace?: boolean
   }
 
   /**
@@ -11626,6 +11667,10 @@ declare module 'coc.nvim' {
     filter?(document: { uri: string, languageId: string }, mode: 'onType' | 'onSave'): boolean
   }
 
+  export interface URIConverter {
+    (value: Uri): string
+  }
+
   export interface LanguageClientOptions {
     ignoredRootPaths?: string[]
     disableSnippetCompletion?: boolean
@@ -11643,6 +11688,10 @@ declare module 'coc.nvim' {
      * to 'utf8' if omitted.
      */
     stdioEncoding?: string
+    // converter used to decode uri.
+    uriConverter?: {
+      code2Protocol: URIConverter
+    }
     initializationOptions?: any | (() => any)
     initializationFailedHandler?: InitializationFailedHandler
     progressOnInitialization?: boolean